The Disassembler octo assembly and octo disassembly now show the register names with their changed name. In case of :alias the :alias is insert at the decompiled Line corrosponding to the source. In case of :struct all Struct members are converted to :alias with Struct and Member name (for example Player.x). The variables also appear as Player.x even within a with clause.
Some work on the samples. Needs to be cleaned up and documented stil
show the register names with their changed name. In case of :alias the
:alias is insert at the decompiled Line corrosponding to the source.
In case of :struct all Struct members are converted to :alias with
Struct and Member name (for example Player.x). The variables also appear
as Player.x even within a with clause.

Some work on the samples. Needs to be cleaned up and documented stil..